Replace "unsigned long" for array indices by uint32_t
One sould not have a type of variable size for implementing array size that are
found in the .gen interface. Let's decide that genom arrays and fixed length
sequences are limited to 2^32-1 elements (which was the case on i386 anyway,
but not on x86_64 so far).

+++ src/declarator.c
@@ -39,7 +39,7 @@ struct dcl_s {
   const char *name;
 
   unsigned int dims;
-  unsigned long dim[0];
+  uint32_t dim[0];
 
   idltype_s type;
 };
@@ -93,7 +93,7 @@ dcl_destroy(dcl_s d)
 /** Add a new dimension to a declarator
  */
 dcl_s
-dcl_adddim(dcl_s dcl, unsigned long dim)
+dcl_adddim(dcl_s dcl, uint32_t dim)
 {
   dcl_s d;
   assert(dcl);
